feat: Majority Judgment Polls (2020)
Add `Poll` model and `CreatePoll` function.

Add `Judgment` creation and update.

There can be at most one judgment per trio of poll, judge and candidate.
Each Judge may emit one Judgment per Candidate of a Poll.

Mop up and oust a reminiscent bug.

Add the Poll Create, Update and Delete forms and logic.

We can't tally a Poll yet.
Nor can we judge issues.

Add the Judgment forms for each Poll of each issue.

Everything works so far without javascript.
We added CSS like monkeys, until we figure out what to do with it.

Use emotes Chromium can print as well.

Fix some warnings.

Prepare for the far future.

Add scoring pseudo-code for fast deliberation.

This was entertaining, but it's perhaps wrong,
although we can't find any reason why so far.
We contacted the world's top specialist on the matter.
…

Three gods, A, B, and C are called,
in no particular order, True, False, and Random.

True always speaks truly, False always speaks falsely,
but whether Random speaks truly or falsely is a
completely random matter.

Your task is to determine the identities of A, B, and C
by asking three yes-no questions; each question
must be put to exactly one god. The gods understand
English, but will answer all questions in
their own language, in which the words for yes
and no are da and ja, in some order.

You do not know which word means which.

Add a naive tallier.

I wish I had a feature suite in Gherkin here…

Add some tests to the deliberator and implement them.

Still waiting for input from the specialist(s)
before delving into implementation details.

Use unsigned integers where it makes sense.

(or not)

Provision more tests for the deliberator.

We've got some answers — still confused, though.

Implement a very naive deliberator.

Fix a fix for the label's hover area.

Let's barf as little as possible on existing CSS.

Fill the missing judgments as REJECT judgments.

This should be optional.

Add a battery of tests to GetMedian

We're now going to be able to implement it properly,
not like the cheap cheat we had so far.

Adapt label color to median grade.

This is a cheap implementation that will probably need a rewrite
for complete support of other gradations.

Add more tests to expose the limitations of the current scoring implementation.

Now we can work some more on the scoring.

Implement a less naive score calculus.

So far it holds.
The score is a bit long (90chars = 15 chars per possible grade)
but it won't get any longer with more judges,
unless we have to go over 500 billion judges.

Fix poll's CRUD subheaders I18N.

Make sure judgments can be emitted with keyboard only.

`:focus-within` <3 U CSS

Add a basic page to view a poll.

Move Poll CSS to LESS.

Clean up.

Tease about what will come next.

Add a repo header nav item for polls.

Ideally this whole thing should be a plugin and not a fork.
Not sure how to add new Db models and Routes.
We'd also need a custom spot in the issue view template,
but that can probably be merged upstream.

Clean up.

We'd love to move the routes to their own file.
And then being able to add them via `custom/`.

Disable event sourcing until /user/events do not hang for a minute,
because it makes browsing gitea rapidly impossible.

Add french translations for the polling UI.

Rename `Judgment` into `PollJudgment`.

After careful consideration, it's probably best to namespace.
Also, explicit is better than implicit.   :)

BREAKING CHANGE:  Not sure how the migrations will operate here.
This is why it's best to do it now than once we're actually using it.

Tweak the poll's labels for mobile with a CSS media query.

Thanks @pierre-louis for the bug report !

docs: explicit is better than implicit

chore: "missing" is more explicit than "lost"

feat: add radial merit profiles, in SVG

Ignore the gitea-repositories directory.

chore: clean up the radial merit profile SVG template

test: expose an issue with the tally

The returned amounts for each grade are wrong.
How could this happen, you ask?
Lack of unit-testing, that's why!
(also, shallow copy shenanigans)

chore: spacing

Still experimenting with the spacing in templates.
I know Gitea does not use spaces.
My sight is not getting any better,
and spacing helps a lot.

Yes, I'm already using bigger fonts.

test: refine the tests about the critical bug found

… so that it does not happen again!

fix: elbow grease the copy mechanism to fix The Bug®

fix: hide the description if none is provided.

Should we mention that it is a bad practice ?
It feels like it is.  … To deliberate.

feat: add the issue title to the list of issues of a poll

Also naive and inefficient, but if we paginate, we'll be ok.
And if not, we can probably improve this by batching the queries.

fix: remove a debugging tweak that was overlooked

Also, remove the need to translate the word `Issue`.
Oddly enough, it is never translated by itself,
it is only in plural form opr inside a sentence.

feat: add the merit profile behind the median grade

Usual CSS woes
- Really not confident about positioning top and left with `em`.
- Z-index shenanigans to skip unexpected offset in positioning,
  where the absolute of a child in a relative container
  won't point to top left but to bottom of other child in static
  (try moving the merit profile below the input.emote, and remove z-indices)

Right now the merit profile is shown to all logged in, whether they judged or not.  This is not the intended final behavior.  Ideally we'd have settings fdr this.

chore: bump brotli since previous version was marked as insecure

fix: update to 1.14.3

fix: issue with the biggest group calculus

I've known about this one for almost a year.
Feels good to finally get around to it.